Cleanups: removing extraneous evaluators created for debugging of transient sensititivites.
