--- src/evaluators/pde/PHAL_SteadyThermalResidWithSensitivities.cpp
@@ -1,13 +1,0 @@
-//*****************************************************************//
-//    Albany 3.0:  Copyright 2016 Sandia Corporation               //
-//    This Software is released under the BSD license detailed     //
-//    in the file "license.txt" in the top-level Albany directory  //
-//*****************************************************************//
-
-#include "PHAL_AlbanyTraits.hpp"
-
-#include "PHAL_SteadyThermalResidWithSensitivities.hpp"
-#include "PHAL_SteadyThermalResidWithSensitivities_Def.hpp"
-
-PHAL_INSTANTIATE_TEMPLATE_CLASS(PHAL::SteadyThermalResidWithSensitivities)
-
