+++ src/problems/Albany_ThermalProblem.hpp
@@ -18,7 +18,6 @@
 
 #include "PHAL_ConvertFieldType.hpp"
 #include "Albany_MaterialDatabase.hpp"
-
 namespace Albany {
 
   /*!
@@ -115,8 +114,9 @@ namespace Albany {
 #include "Albany_ResponseUtilities.hpp"
 //#include "PHAL_Neumann.hpp"
 #include "PHAL_ThermalResid.hpp"
-#include "PHAL_ThermalSource.hpp"
-
+//IKT FIXME - add dependence on LandIce
+#include "../../LandIce/evaluators/LandIce_SharedParameter.hpp"
+#include "../../LandIce/problems/LandIce_ParamEnum.hpp"
 
 template <typename EvalT>
 Teuchos::RCP<const PHX::FieldTag>
@@ -205,24 +205,15 @@ Albany::ThermalProblem::constructEvaluators(
         evalUtils.constructDOFGradInterpolationEvaluator(dof_names[i]));
   }
 
-  {  // Temperature Source
-    RCP<ParameterList> p = rcp(new ParameterList("Temperature Resid"));
-    p->set<RCP<DataLayout>>("Node QP Scalar Data Layout", dl->node_qp_scalar);
-    p->set<RCP<DataLayout>>("QP Scalar Data Layout", dl->qp_scalar);
-    p->set<RCP<DataLayout>>("QP Vector Data Layout", dl->qp_vector);
-    p->set<string>("QP Coordinate Vector Name", "Coord Vec");
-    
-    p->set<Teuchos::Array<double>>("Thermal Conductivity", kappa);
-    p->set<double>("Heat Capacity", C);
-    p->set<double>("Density", rho);
-    p->set<std::string>("Thermal Source", thermal_source); 
-    
-    // Output
-    p->set<string>("Source Name", "Temperature Source");
-    p->set<RCP<DataLayout>>("Node Scalar Data Layout", dl->node_scalar);
-
-    ev = rcp(new PHAL::ThermalSource<EvalT, AlbanyTraits>(*p));
-    fm0.template registerEvaluator<EvalT>(ev);
+  {  //Shared parameters for sensitivity analysis 
+    RCP<ParameterList> p = rcp(new ParameterList("Thermal Conductivity Parameters"));
+    p->set< RCP<ParamLib> >("Parameter Library", paramLib);
+    const std::string param_name = "kappa_x Parameter";
+    p->set<std::string>("Parameter Name", param_name);
+    RCP<LandIce::SharedParameter<EvalT,PHAL::AlbanyTraits,LandIce::ParamEnum,LandIce::ParamEnum::Homotopy>> ptr_kappa_x;
+    ptr_kappa_x = rcp(new LandIce::SharedParameter<EvalT,PHAL::AlbanyTraits,LandIce::ParamEnum,LandIce::ParamEnum::Homotopy>(*p,dl));
+    ptr_kappa_x->setNominalValue(params->sublist("Parameters"), kappa[0]); 
+    fm0.template registerEvaluator<EvalT>(ptr_kappa_x);
   }
 
   {  // Temperature Resid
@@ -234,21 +225,25 @@ Albany::ThermalProblem::constructEvaluators(
     p->set<string>("Gradient QP Variable Name", "Temperature Gradient");
     p->set<string>("Weighted Gradient BF Name", "wGrad BF");
     p->set<string>("Source Name", "Temperature Source");
+    p->set<string>("QP Coordinate Vector Name", "Coord Vec");
+    p->set<string>("Source Name", "Temperature Source");
 
     p->set<RCP<DataLayout>>("Node QP Scalar Data Layout", dl->node_qp_scalar);
     p->set<RCP<DataLayout>>("QP Scalar Data Layout", dl->qp_scalar);
     p->set<RCP<DataLayout>>("QP Vector Data Layout", dl->qp_vector);
     p->set<RCP<DataLayout>>("Node QP Vector Data Layout", dl->node_qp_vector);
-
-    p->set<Teuchos::Array<double>>("Thermal Conductivity", kappa);
+    p->set<RCP<DataLayout>>("Node Scalar Data Layout", dl->node_scalar);
+    p->set<std::string>("Continuation Parameter Name","kappa_x Parameter");
     p->set<double>("Heat Capacity", C);
     p->set<double>("Density", rho);
+    p->set<double>("kappa_y", kappa[1]);
+    p->set<double>("kappa_z", kappa[2]);
+    p->set<std::string>("Thermal Source", thermal_source); 
 
     // Output
     p->set<string>("Residual Name", "Temperature Residual");
-    p->set<RCP<DataLayout>>("Node Scalar Data Layout", dl->node_scalar);
 
-    ev = rcp(new PHAL::ThermalResid<EvalT, AlbanyTraits>(*p));
+    ev = rcp(new PHAL::ThermalResid<EvalT, AlbanyTraits>(*p, dl));
     fm0.template registerEvaluator<EvalT>(ev);
   }
