--- tests/small/Thermal1D/input_steady.yaml
@@ -1,88 +1,0 @@
-ALBANY:
-  Debug Output: 
-    Write Solution to MatrixMarket: -1
-  Build Type: Tpetra
-  Problem: 
-    Name: Steady Thermal With Sensitivities 1D
-    Solution Method: Steady
-    Compute Sensitivities: true
-    Dirichlet BCs: 
-      SDBC on NS NodeSet0 for DOF T: 0.0
-      SDBC on NS NodeSet1 for DOF T: 50.0
-    Thermal Conductivity: [1.6] 
-    Thermal Source: None 
-    Response Functions: 
-      Number: 1
-      Response 0: Solution Average
-    Parameters: 
-      Number: 1
-      Parameter 0: 'kappa_x'
-  Discretization: 
-    1D Elements: 20
-    1D Scale: 1.00000000000000000e+00
-    Workset Size: -1
-    Method: STK1D
-    Exodus Output File Name: thermal1D_with_source_out.exo
-  Regression Results: 
-    Number of Comparisons: 1
-    Test Values: [2.500000000000e+01]
-    #Sensitivity Comparisons 0: 
-    #  Number of Sensitivity Comparisons: 1
-    #  Sensitivity Test Values 0: [5.62650000000000025e-02]
-    Relative Tolerance: 1.00000000000000002e-03
-    Absolute Tolerance: 1.00000000000000008e-05
-  Piro: 
-    Sensitivity Method: Forward
-    NOX: 
-      Direction: 
-        Method: Newton
-        Newton: 
-          Forcing Term Method: Constant
-          Rescue Bad Newton Solve: true
-          Stratimikos Linear Solver: 
-            NOX Stratimikos Options: { }
-            Stratimikos: 
-              Linear Solver Type: Belos
-              Linear Solver Types: 
-                AztecOO: 
-                  Forward Solve: 
-                    AztecOO Settings: 
-                      Aztec Solver: GMRES
-                      Convergence Test: r0
-                      Size of Krylov Subspace: 200
-                      Output Frequency: 10
-                    Max Iterations: 200
-                    Tolerance: 1.00000000000000008e-05
-                Belos: 
-                  Solver Type: Block GMRES
-                  Solver Types: 
-                    Block GMRES: 
-                      Convergence Tolerance: 1.00000000000000008e-05
-                      Output Frequency: 10
-                      Output Style: 1
-                      Verbosity: 33
-                      Maximum Iterations: 100
-                      Block Size: 1
-                      Num Blocks: 50
-                      Flexible Gmres: false
-              Preconditioner Type: Ifpack2
-              Preconditioner Types: 
-                Ifpack2: 
-                  Overlap: 2
-                  Prec Type: ILUT
-                  Ifpack2 Settings: 
-                    'fact: drop tolerance': 0.00000000000000000e+00
-                    'fact: ilut level-of-fill': 1.00000000000000000e+00
-                    'fact: level-of-fill': 2
-      Line Search: 
-        Full Step: 
-          Full Step: 1.00000000000000000e+00
-        Method: Full Step
-      Nonlinear Solver: Line Search Based
-      Printing: 
-        Output Information: 103
-        Output Precision: 3
-        Output Processor: 0
-      Solver Options: 
-        Status Test Check Type: Minimal
-...
