+++ tests/small/Thermal1D/input_with_source.yaml
@@ -2,7 +2,7 @@ ALBANY:
   Debug Output: 
     Write Solution to MatrixMarket: 0
   Problem: 
-    Name: Thermal With Sensitivities 1D
+    Name: Thermal 1D
     Solution Method: Transient
     Compute Sensitivities: true
     Dirichlet BCs:
